Reduce the extracted dartdoc data for large packages.
