+++ app/lib/dartdoc/dartdoc_runner.dart
@@ -23,7 +23,6 @@ import '../shared/urls.dart';
 import '../shared/versions.dart' as versions;
 
 import 'backend.dart';
-import 'coverage.dart';
 import 'customization.dart';
 import 'models.dart';
 
@@ -244,7 +243,7 @@ class DartdocJobProcessor extends JobProcessor {
       await toolEnvRef.release();
     }
 
-    final coverage = calculateCoverage(dartdocData);
+    final coverage = dartdocData.coverage ?? Coverage(total: 0, documented: 0);
 
     if (hasContent && dartdocData != null) {
       if (coverage.penalty > 0) {
